Revert "Revert "https://inside.tensor.ru/opendoc.html?guid=c3362ff8-4a31-4caf-a284-c0832c4ac4d5&des=""
This reverts commit cc3df38600fa0044271996e70fdbaee6899a7501.

+++ components/RichEditor/RichEditorComponents/RichTextArea/RichTextArea.module.js
@@ -15,7 +15,7 @@ define('js!SBIS3.CONTROLS.RichTextArea',
    "html!SBIS3.CONTROLS.RichTextArea",
    "js!SBIS3.CONTROLS.Utils.RichTextAreaUtil",
    "js!SBIS3.CONTROLS.RichTextArea/resources/smiles",
-   "js!SBIS3.Plugin.Source.LocalService",
+   'js!WS.Data/Di',
    "js!SBIS3.CONTROLS.Utils.ImageUtil",
    "Core/Sanitize",
    "Core/helpers/collection-helpers",
@@ -27,7 +27,7 @@ define('js!SBIS3.CONTROLS.RichTextArea',
    "css!SBIS3.CORE.RichContentStyles",
    "i18n!SBIS3.CONTROLS.RichEditor",
    'css!SBIS3.CONTROLS.RichTextArea'
-], function( UserConfig, cPathResolver, cContext, cIndicator, cFunctions, CommandDispatcher, cConstants, Deferred,TextBoxBase, dotTplFn, RichUtil, smiles, LocalService, ImageUtil, Sanitize, colHelpers, fcHelpers, strHelpers, dcHelpers, ImageOptionsPanel, EventBus) {
+], function( UserConfig, cPathResolver, cContext, cIndicator, cFunctions, CommandDispatcher, cConstants, Deferred,TextBoxBase, dotTplFn, RichUtil, smiles, Di, ImageUtil, Sanitize, colHelpers, fcHelpers, strHelpers, dcHelpers, ImageOptionsPanel, EventBus) {
       'use strict';
       //TODO: ПЕРЕПИСАТЬ НА НОРМАЛЬНЫЙ КОД РАБОТУ С ИЗОБРАЖЕНИЯМИ
       var
@@ -465,15 +465,9 @@ define('js!SBIS3.CONTROLS.RichTextArea',
                   event.preventDefault();
                   return false;
                },
-               //service создаётся каждый раз и destroy`тся каждый раз тк плагин может перезагрузиться и сервис протухнет
-               //см прохождение по задаче:https://inside.tensor.ru/opendoc.html?guid=c3362ff8-4a31-4caf-a284-c0832c4ac4d5&des=
-               service = new LocalService({
-                  endpoint: {
-                     address: 'Clipboard-1.0.1.0',
-                     contract: 'Clipboard'
-                  },
-                  options: { mode: 'silent' }
-               }),
+               service = {
+                  destroy: function(){}
+               },
                createDialog = function() {
                   cIndicator.hide();
                   require(['js!SBIS3.CORE.Dialog', 'js!SBIS3.CONTROLS.Button'], function(Dialog, Button) {
@@ -519,25 +513,37 @@ define('js!SBIS3.CONTROLS.RichTextArea',
                };
 
             cIndicator.show();
-
-            service.isReady().addCallback(function() {
-               service.call("getContentType", {}).addCallback(function (ContentType) {
-                  service.call(ContentType === 'Text/Html' || ContentType === 'Text/Rtf' || ContentType === 'Html' || ContentType === 'Rtf' ? 'getHtml' : 'getText', {}).addCallback(function (content) {
-                     cIndicator.hide();
-                     prepareAndInsertContent(content);
-                     if (typeof onAfterCloseHandler === 'function') {
-                        onAfterCloseHandler();
-                     }
-                     service.destroy();
+            if (Di.isRegistered('SBIS3.Plugin.Source.LocalService')) {
+               //service создаётся каждый раз и destroy`тся каждый раз тк плагин может перезагрузиться и сервис протухнет
+               //см прохождение по задаче:https://inside.tensor.ru/opendoc.html?guid=c3362ff8-4a31-4caf-a284-c0832c4ac4d5&des=
+               service = Di.resolve('SBIS3.Plugin.Source.LocalService',{
+                  endpoint: {
+                     address: 'Clipboard-1.0.1.0',
+                     contract: 'Clipboard'
+                  },
+                  options: { mode: 'silent' }
+               });
+               service.isReady().addCallback(function() {
+                  service.call("getContentType", {}).addCallback(function (ContentType) {
+                     service.call(ContentType === 'Text/Html' || ContentType === 'Text/Rtf' || ContentType === 'Html' || ContentType === 'Rtf' ? 'getHtml' : 'getText', {}).addCallback(function (content) {
+                        cIndicator.hide();
+                        prepareAndInsertContent(content);
+                        if (typeof onAfterCloseHandler === 'function') {
+                           onAfterCloseHandler();
+                        }
+                        service.destroy();
+                     }).addErrback(function () {
+                        createDialog();
+                     });
                   }).addErrback(function () {
                      createDialog();
                   });
                }).addErrback(function () {
                   createDialog();
                });
-            }).addErrback(function () {
+            } else {
                createDialog();
-            });
+            }
          },
          /**
           * <wiTag group="Управление">
